refactor: rename story
